loginUpdate
modification de la fonctionnalité login, masquage de ce qui doit être masqué, effacement des variables de session

+++ deleteTeacher.php
@@ -4,7 +4,7 @@
     include_once("Database.php");
     $database = new Database();
 
-    if (array_key_exists("id", $_GET))
+    if (array_key_exists("userPermissionsNumber", $_SESSION) && $_SESSION["userPermissionsNumber"] >= 100 &&  array_key_exists("id", $_GET))
     {
         $database->deleteTeacher($_GET["id"]);
         header('Location: index.php?controller=teacher&action=list&delete=true');
